app/requirements/base.txt
Django==3.1
psycopg2-binary==2.8.6
python-dotenv
djangorestframework==3.12.1
factory_boy==3.0.1
gunicorn
uvicorn[standard]
